mobile_movie_cut/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
jinja2>=3.1.0
sse-starlette>=2.0.0
aiofiles>=24.0.0
python-multipart>=0.0.9
pywebview>=4.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
